requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain_community
langchain_openai
openai
python-dotenv
rapidocr-onnxruntime
unstructured[pptx, md, xlsx]
markdown
pyqt5
PyQt-Fluent-Widgets[full] -i https://pypi.org/simple/